requirements.txt
pydantic==2.10.1
numpy==2.1.3
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
jsonschema==4.23.0

## LLM
openai==1.55.0

## Data
openmeteo-requests==1.3.0